+++ client/web/src/auth/PostSignUpPage.tsx
@@ -9,7 +9,7 @@ import { ErrorLike } from '@sourcegraph/shared/src/util/errors'
 import { BrandLogo } from '@sourcegraph/web/src/components/branding/BrandLogo'
 import { HeroPage } from '@sourcegraph/web/src/components/HeroPage'
 
-import { AuthenticatedUser } from '../auth'
+import { AuthenticatedUser } from '../../../../wildcard'
 import { PageTitle } from '../components/PageTitle'
 import { SourcegraphContext } from '../jscontext'
 import { useTemporarySetting } from '../settings/temporary/useTemporarySetting'
@@ -20,7 +20,7 @@ import { UserExternalServicesOrRepositoriesUpdateProps } from '../util'
 import styles from './PostSignUpPage.module.scss'
 import { getReturnTo } from './SignInSignUpCommon'
 import signInSignUpCommonStyles from './SignInSignUpCommon.module.scss'
-import { Steps, Step, StepList, StepPanels, StepPanel, StepActions } from './Steps'
+import { Steps, Step, StepList, StepPanels, StepPanel, StepActions } from '../../../../wildcard/components'
 import { useExternalServices } from './useExternalServices'
 import { CodeHostsConnection } from './welcome/CodeHostsConnection'
 import { Footer } from './welcome/Footer'
