+++ client/web/src/auth/welcome/StartSearching.tsx
@@ -4,12 +4,12 @@ import React, { useEffect, useState } from 'react'
 import { Link } from '@sourcegraph/shared/src/components/Link'
 import { ErrorLike, isErrorLike } from '@sourcegraph/shared/src/util/errors'
 
-import { AuthenticatedUser } from '../../auth'
+import { AuthenticatedUser } from '../../../../../wildcard'
 import { eventLogger } from '../../tracking/eventLogger'
 import { UserExternalServicesOrRepositoriesUpdateProps } from '../../util'
 import { LogoAscii } from '../LogoAscii'
 import { getPostSignUpEvent, RepoSelectionMode } from '../PostSignUpPage'
-import { useSteps } from '../Steps'
+import { useSteps } from '../../../../../wildcard/components'
 import { Terminal, TerminalTitle, TerminalLine, TerminalDetails, TerminalProgress } from '../Terminal'
 import { useExternalServices } from '../useExternalServices'
 import { useRepoCloningStatus } from '../useRepoCloningStatus'
