+++ client/web/src/user/settings/repositories/SelectAffiliatedRepos.tsx
@@ -8,7 +8,7 @@ import { ErrorLike } from '@sourcegraph/shared/src/util/errors'
 import { Container, PageSelector } from '@sourcegraph/wildcard'
 
 import { RepoSelectionMode } from '../../../auth/PostSignUpPage'
-import { useSteps } from '../../../auth/Steps'
+import { useSteps } from '../../../../../../wildcard/components'
 import { useAffiliatedRepos } from '../../../auth/useAffiliatedRepos'
 import { useExternalServices } from '../../../auth/useExternalServices'
 import { useSelectedRepos, selectedReposVar, MinSelectedRepo } from '../../../auth/useSelectedRepos'
